backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt
jwt
python-dotenv
werkzeug